# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AR.IO Bundler Lite API
  description: |
    Lightweight Arweave ANS-104 bundler with x402 USDC payment integration.

    ## x402 Payment Flow

    1. **Get Price Quote**: `GET /v1/x402/price/{signatureType}/{address}?bytes={size}`
    2. **Create EIP-712 Signature**: Sign payment authorization using wallet
    3. **Upload with Payment**: `POST /v1/tx` with `X-PAYMENT` header containing base64-encoded payment payload
    4. **Finalization** (automatic): Service validates actual byte count and applies fraud detection (±5% tolerance)

    ## Supported Networks

    - **Base Mainnet** (chainId: 8453) - Production USDC payments
    - **Base Sepolia** (chainId: 84532) - Testnet USDC payments
    - **Ethereum Mainnet** (chainId: 1)
    - **Polygon** (chainId: 137)

    ## Authentication

    - **x402 Payment Header**: Base64-encoded JSON containing EIP-712 signature and payment details
    - No API keys required for x402 payments (stateless)
  version: 1.0.0
  contact:
    name: AR.IO Bundler Lite
    url: https://github.com/ar-io/ar-io-bundler
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.example.com
    description: Production server (configure as needed)
...
paths:
  /v1/x402/price/{signatureType}/{address}:
    get:
      tags:
        - x402
      summary: Get price quote for data upload
      description: |
        Returns the USDC amount required to upload data of a given size.
        Uses the configured network's USDC token and current AR/USD pricing.

        **Price Calculation:**
        1. Convert byte count to Winston credits using current AR/USD rate
        2. Apply network fee multiplier
        3. Convert Winston to USDC atomic units (6 decimals)

        **Note:** Price quotes are valid for a short time window. Create payment immediately after receiving quote.
      operationId: getX402Price
      parameters:
        - name: signatureType
          in: path
          required: true
          description: Arweave signature type (always use 3 for x402/EIP-712)
          schema:
            type: integer
            enum: [3]
            example: 3
        - name: address
          in: path
          required: true
          description: User's EVM wallet address (0x-prefixed)
          schema:
            type: string
            pattern: '^0x[a-fA-F0-9]{40}$'
            example: '0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb'
        - name: bytes
          in: query
          required: true
          description: Number of bytes to upload
          schema:
            type: integer
            minimum: 1
            maximum: 10737418240
            example: 1024
        - name: network
          in: query
          required: false
          description: EVM network for payment (defaults to base-sepolia for testnet)
          schema:
            type: string
            enum: [base, base-sepolia, ethereum, polygon]
            default: base-sepolia
            example: base-sepolia
      responses:
        '200':
          description: Successful price quote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/X402PriceQuote'
              example:
                usdcAmount: '50000'
                wincAmount: '1000000000'
                network: base-sepolia
                chainId: 84532
                usdcAddress: '0x036CbD53842c5426634e7929541eC2318f3dCF7e'
                paymentAddress: '0xYourBundlerAddress'
                byteCount: 1024
                expiresAt: '2024-11-04T20:30:00Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/x402/payment/{signatureType}/{address}:
    post:
      tags:
        - x402
      summary: Verify x402 payment authorization
      description: |
        Verifies an EIP-712 signed payment authorization for data upload.
        This endpoint validates the signature, checks USDC allowance/balance,
        and returns a payment ID to include in the upload request.

        **Payment Verification:**
        1. Validates EIP-712 signature against USDC contract domain
        2. Checks that `from` address has sufficient USDC balance
        3. Verifies USDC allowance for transferWithAuthorization
        4. Checks that authorization hasn't been used (nonce replay protection)
        5. Creates payment record in database

        **Important:** This does NOT transfer funds. Actual USDC transfer occurs
        after upload is finalized and fraud detection passes.
      operationId: verifyX402Payment
      parameters:
        - name: signatureType
          in: path
          required: true
          schema:
            type: integer
            enum: [3]
            example: 3
        - name: address
          in: path
          required: true
          description: User's EVM wallet address (must match signature)
          schema:
            type: string
            pattern: '^0x[a-fA-F0-9]{40}$'
            example: '0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/X402PaymentRequest'
            example:
              transferParams:
                from: '0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb'
                to: '0xYourBundlerAddress'
                value: '50000'
                validAfter: 0
                validBefore: 1730754000
                nonce: '0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef'
              signature: '0xabcdef...'
              network: base-sepolia
              byteCount: 1024
      responses:
        '200':
          description: Payment verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/X402PaymentResponse'
              example:
                paymentId: 'pmt_1234567890abcdef'
                status: 'pending_validation'
                wincAmount: '1000000000'
                dataItemId: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/x402/finalize:
    post:
      tags:
        - x402
      summary: Finalize x402 payment with fraud detection
      description: |
        **Internal endpoint** - Called automatically after data upload completes.

        Applies fraud detection by comparing declared vs actual byte count:
        - **Within ±5% tolerance**: Payment confirmed, USDC transferred
        - **Under-declared (>5%)**: Fraud penalty, full payment charged
        - **Over-declared (>5%)**: Refund issued for excess amount

        This endpoint triggers the actual USDC transfer using Coinbase CDP API
        (mainnet) or on-chain settlement (testnet).
      operationId: finalizeX402Payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - dataItemId
                - actualByteCount
              properties:
                dataItemId:
                  type: string
                  description: Data item transaction ID
                  example: 'abcdef1234567890abcdef1234567890abcdef1234567890'
                actualByteCount:
                  type: integer
                  minimum: 1
                  description: Actual number of bytes uploaded
                  example: 1050
      responses:
        '200':
          description: Payment finalized successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [confirmed, refunded, fraud_penalty]
                    description: Final payment status
                  refundWinc:
                    type: string
                    description: Refund amount in Winston (if over-declared)
                    nullable: true
                  message:
                    type: string
                    description: Human-readable status message
              example:
                status: 'confirmed'
                refundWinc: null
                message: 'Payment confirmed within tolerance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /price/x402/data-item/{token}/{byteCount}:
    get:
      tags:
        - x402
      summary: Get price for signed data item (Turbo-style)
      description: |
        Returns x402 payment requirements for uploading a complete, signed ANS-104 data item.

        **Use this endpoint when you:**
        - Already have a signed ANS-104 data item
        - Know the exact byte count of your data item

        **Token Format:** `{currency}-{network}`
        - `usdc-base` - USDC on Base mainnet
        - `usdc-base-sepolia` - USDC on Base Sepolia testnet
        - `usdc-ethereum-mainnet` - USDC on Ethereum
        - `usdc-polygon-mainnet` - USDC on Polygon

        **Pricing:**
        1. Byte count → Winston cost (via Arweave pricing)
        2. Add 30% bundler fee (profit margin on top of Arweave costs)
        3. Winston → USDC atomic units (via CoinGecko AR/USD rate)
        4. Apply minimum threshold (0.001 USDC = 1,000 atomic units)

        Returns standard x402 v1 payment requirements.
      operationId: getX402DataItemPrice
      parameters:
        - name: token
          in: path
          required: true
          description: Payment token in format {currency}-{network}
          schema:
            type: string
            enum: [usdc-base, usdc-base-sepolia, usdc-ethereum-mainnet, usdc-polygon-mainnet]
            example: usdc-base
        - name: byteCount
          in: path
          required: true
          description: Exact byte count of signed data item
          schema:
            type: integer
            minimum: 1
            maximum: 10737418240
            example: 2048
      responses:
        '200':
          description: x402 payment requirements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/X402PaymentRequiredResponse'
              example:
                x402Version: 1
                accepts:
                  - scheme: exact
                    network: base
                    maxAmountRequired: "1234"
                    resource: "http://localhost:3001/v1/tx"
                    description: "Upload data to Arweave via AR.IO Bundler"
                    payTo: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
                    maxTimeoutSeconds: 300
                    asset: "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913"
                    extra:
                      name: "USD Coin"
                      version: "2"
        '400':
          description: Invalid token or byte count
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  token:
                    type: string
                  supported:
                    type: array
                    items:
                      type: string
              example:
                error: "Invalid or unsupported token"
                token: "invalid-token"
                supported: ["usdc-base", "usdc-base-sepolia"]
        '500':
          $ref: '#/components/responses/InternalError'
  /price/x402/data/{token}/{byteCount}:
    get:
      tags:
        - x402
      summary: Get price for raw data (bundler creates data item)
      description: |
        Returns x402 payment requirements for uploading raw data where the bundler
        will create the ANS-104 data item wrapper.

        **Use this endpoint when you:**
        - Have raw data (not a signed data item)
        - Want the bundler to create and sign the data item
        - Need pricing that includes ANS-104 wrapper overhead

        **Overhead Estimation:**
        The final data item size is estimated by adding:
        - Raw data bytes (your data)
        - ANS-104 wrapper (~1.1 KB for signature, owner, headers)
        - User tags (if provided via `tags` parameter)
        - Content-Type tag (if provided via `contentType` parameter)
        - System tags (7 x402 metadata tags)

        **System Tags Added:**
        1. Bundler
        2. Upload-Type
        3. Payer-Address
        4. X402-TX-Hash
        5. X402-Payment-ID
        6. X402-Network
        7. Upload-Timestamp

        **Response includes `sizeBreakdown`** showing detailed size calculation.
      operationId: getX402RawDataPrice
      parameters:
        - name: token
          in: path
          required: true
          description: Payment token in format {currency}-{network}
          schema:
            type: string
            enum: [usdc-base, usdc-base-sepolia, usdc-ethereum-mainnet, usdc-polygon-mainnet]
            example: usdc-base
        - name: byteCount
          in: path
          required: true
          description: Raw data byte count (before ANS-104 wrapping)
          schema:
            type: integer
            minimum: 1
            maximum: 10737418240
            example: 1024
        - name: tags
          in: query
          required: false
          description: Number of user tags to include
          schema:
            type: integer
            minimum: 0
            maximum: 100
            default: 0
            example: 3
        - name: contentType
          in: query
          required: false
          description: MIME type (adds Content-Type tag if provided)
          schema:
            type: string
            example: "image/png"
      responses:
        '200':
          description: x402 payment requirements with size breakdown
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/X402PaymentRequiredResponse'
                  - type: object
                    properties:
                      sizeBreakdown:
                        type: object
                        properties:
                          rawDataBytes:
                            type: integer
                            description: Original raw data size
                          ans104Overhead:
                            type: integer
                            description: ANS-104 wrapper + tags overhead
                          estimatedTotalBytes:
                            type: integer
                            description: Final estimated data item size
                          tagCount:
                            type: object
                            properties:
                              user:
                                type: integer
                              contentType:
                                type: integer
                              system:
                                type: integer
                              total:
                                type: integer
              example:
                x402Version: 1
                sizeBreakdown:
                  rawDataBytes: 1024
                  ans104Overhead: 1317
                  estimatedTotalBytes: 2341
                  tagCount:
                    user: 3
                    contentType: 1
                    system: 7
                    total: 11
                accepts:
                  - scheme: exact
                    network: base
                    maxAmountRequired: "1456"
                    resource: "http://localhost:3001/v1/tx"
                    description: "Upload data to Arweave via AR.IO Bundler"
                    payTo: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
                    maxTimeoutSeconds: 300
                    asset: "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913"
        '400':
          description: Invalid token, byte count, or tags parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
              example:
                error: "Invalid tags parameter"
                message: "tags must be a non-negative integer"
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/price/x402/data-item/{token}/{byteCount}:
    get:
      tags:
        - x402
      summary: Get price for signed data item (with /v1 prefix)
      description: |
        **Identical to `/price/x402/data-item/{token}/{byteCount}`**

        This endpoint is the same as the non-prefixed version, provided for consistency.
        See `/price/x402/data-item/{token}/{byteCount}` for full documentation.
      operationId: getX402DataItemPriceV1
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            enum: [usdc-base, usdc-base-sepolia, usdc-ethereum-mainnet, usdc-polygon-mainnet]
        - name: byteCount
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 10737418240
      responses:
        '200':
          description: x402 payment requirements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/X402PaymentRequiredResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/price/x402/data/{token}/{byteCount}:
    get:
      tags:
        - x402
      summary: Get price for raw data (with /v1 prefix)
      description: |
        **Identical to `/price/x402/data/{token}/{byteCount}`**

        This endpoint is the same as the non-prefixed version, provided for consistency.
        See `/price/x402/data/{token}/{byteCount}` for full documentation.
      operationId: getX402RawDataPriceV1
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            enum: [usdc-base, usdc-base-sepolia, usdc-ethereum-mainnet, usdc-polygon-mainnet]
        - name: byteCount
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 10737418240
        - name: tags
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: contentType
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: x402 payment requirements with size breakdown
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/X402PaymentRequiredResponse'
                  - type: object
                    properties:
                      sizeBreakdown:
                        type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/tx:
    post:
      tags:
        - upload
      summary: Upload signed data item
      description: |
        Upload a signed ANS-104 data item with x402 payment.

        **Flow:**
        1. Get price quote from `/v1/x402/price`
        2. Create EIP-712 signature for payment
        3. Verify payment at `/v1/x402/payment`
        4. Upload data with `X-PAYMENT` header containing base64-encoded payment JSON

        **Payment Header Format:**
        ```json
        {
          "paymentId": "pmt_xxx",
          "signature": "0x...",
          "transferParams": {...},
          "network": "base-sepolia"
        }
        ```

        The service will:
        - Verify payment authorization
        - Accept data item for bundling
        - Queue async jobs (bundling, posting to Arweave)
        - Finalize payment with fraud detection
      operationId: uploadDataItem
      parameters:
        - name: X-PAYMENT
          in: header
          required: true
          description: Base64-encoded x402 payment payload
          schema:
            type: string
            format: byte
            example: 'eyJwYXltZW50SWQiOiJwbXRfMTIzNCIsInNpZ25hdHVyZSI6IjB4Li4uIn0='
      requestBody:
        required: true
        description: Signed ANS-104 data item binary
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Data item accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
              example:
                id: 'abcdef1234567890abcdef1234567890abcdef1234567890'
                dataCaches: ['arweave.net']
                fastFinalityIndexes: []
                owner: '0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb'
                signature: '0x...'
                deadlineHeight: 1000000
                block: 950000
                validatorSignatures: []
                timestamp: 1730750000000
                version: '1.0.0'
                public: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '413':
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: 'Data item exceeds maximum size of 10GB'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/tx/{id}/status:
    get:
      tags:
        - status
      summary: Get data item status
      description: |
        Returns the current status of an uploaded data item.

        **Status Values:**
        - `pending`: Uploaded, waiting to be bundled
        - `bundled`: Included in a bundle, waiting to be posted
        - `posted`: Posted to Arweave, waiting for confirmation
        - `permanent`: Confirmed on Arweave blockchain
        - `failed`: Upload or bundling failed
      operationId: getDataItemStatus
      parameters:
        - name: id
          in: path
          required: true
          description: Data item transaction ID
          schema:
            type: string
            minLength: 43
            maxLength: 43
            example: 'abcdef1234567890abcdef1234567890abcdef1234567890'
      responses:
        '200':
          description: Status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataItemStatus'
              example:
                id: 'abcdef1234567890abcdef1234567890abcdef1234567890'
                status: 'permanent'
                bundleId: 'bundle_1234567890abcdef'
                blockHeight: 1000500
                blockTimestamp: 1730760000
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/tx/{id}/offsets:
    get:
      tags:
        - status
      summary: Get data item offset information
      description: |
        Returns offset information for retrieving the data item from Arweave.
        Only available after the data item has been posted to Arweave.
      operationId: getDataItemOffsets
      parameters:
        - name: id
          in: path
          required: true
          description: Data item transaction ID
          schema:
            type: string
            example: 'abcdef1234567890abcdef1234567890abcdef1234567890'
      responses:
        '200':
          description: Offset information retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataItemOffset'
              example:
                id: 'abcdef1234567890abcdef1234567890abcdef1234567890'
                offset: 123456789
                size: 1024
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
    X402PriceQuote:
      type: object
      required:
        - usdcAmount
        - wincAmount
        - network
        - chainId
        - usdcAddress
        - paymentAddress
        - byteCount
        - expiresAt
      properties:
        usdcAmount:
          type: string
          description: USDC amount in atomic units (6 decimals) as string
          example: '50000'
        wincAmount:
          type: string
          description: Equivalent Winston credits as string
          example: '1000000000'
        network:
          type: string
          description: EVM network identifier
          enum: [base, base-sepolia, ethereum, polygon]
          example: 'base-sepolia'
        chainId:
          type: integer
          description: EVM chain ID
          example: 84532
        usdcAddress:
          type: string
          description: USDC token contract address
          pattern: '^0x[a-fA-F0-9]{40}$'
          example: '0x036CbD53842c5426634e7929541eC2318f3dCF7e'
        paymentAddress:
          type: string
          description: Bundler's payment receiver address
          pattern: '^0x[a-fA-F0-9]{40}$'
          example: '0xYourBundlerAddress'
        byteCount:
          type: integer
          description: Number of bytes quoted
          example: 1024
        expiresAt:
          type: string
          format: date-time
          description: Quote expiration timestamp (ISO 8601)
          example: '2024-11-04T20:30:00Z'
    X402PaymentRequest:
      type: object
      required:
        - transferParams
        - signature
        - network
      properties:
        transferParams:
          $ref: '#/components/schemas/TransferParams'
        signature:
          type: string
          description: EIP-712 signature (0x-prefixed hex)
          pattern: '^0x[a-fA-F0-9]{130}$'
          example: '0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890ab'
        network:
          type: string
          enum: [base, base-sepolia, ethereum, polygon]
          example: 'base-sepolia'
        byteCount:
          type: integer
          minimum: 1
          description: Declared byte count for fraud detection
          example: 1024
...
    X402PaymentResponse:
      type: object
      required:
        - paymentId
        - status
        - wincAmount
      properties:
        paymentId:
          type: string
          description: Unique payment identifier
          example: 'pmt_1234567890abcdef'
        status:
          type: string
          enum: [pending_validation, confirmed, refunded, fraud_penalty]
          description: Current payment status
          example: 'pending_validation'
        wincAmount:
          type: string
          description: Winston credits allocated
          example: '1000000000'
        dataItemId:
          type: string
          nullable: true
          description: Associated data item ID (null until upload)
          example: null
    UploadResponse:
      type: object
      required:
        - id
        - timestamp
      properties:
        id:
          type: string
          description: Data item transaction ID
          example: 'abcdef1234567890abcdef1234567890abcdef1234567890'
        dataCaches:
          type: array
          items:
            type: string
          description: List of caches that will index this data
          example: ['arweave.net']
        fastFinalityIndexes:
          type: array
          items:
            type: string
          description: Fast finality indexes
          example: []
        owner:
          type: string
          description: Owner address
          example: '0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb'
        signature:
          type: string
          description: Data item signature
          example: '0x...'
        deadlineHeight:
          type: integer
          description: Arweave block height deadline
          example: 1000000
        block:
          type: integer
          description: Current Arweave block height
          example: 950000
        validatorSignatures:
          type: array
          items:
            type: string
          example: []
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp (milliseconds)
          example: 1730750000000
        version:
          type: string
          description: API version
          example: '1.0.0'
        public:
          type: string
          nullable: true
          description: Public key (if applicable)
          example: null
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Invalid byte count parameter'
            code: 'BAD_REQUEST'
    PaymentRequired:
      description: Payment required - x402 payment failed or insufficient
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Invalid payment signature'
            code: 'PAYMENT_REQUIRED'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Data item not found'
            code: 'NOT_FOUND'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Internal server error'
            code: 'INTERNAL_ERROR'
  securitySchemes:
    X402Payment:
      type: apiKey
      in: header
      name: X-PAYMENT
      description: |
        Base64-encoded JSON containing x402 payment authorization.

        **Structure:**
        ```json
        {
          "paymentId": "pmt_xxx",
          "signature": "0x...",
          "transferParams": {
            "from": "0x...",
            "to": "0x...",
            "value": "50000",
            "validAfter": 0,
            "validBefore": 1730754000,
            "nonce": "0x..."
          },
          "network": "base-sepolia"
        }
        ```
security:
  - X402Payment: []